+++ packages/vest/src/hooks/focused/__tests__/focused.test.ts
@@ -1,602 +1,466 @@
-import { faker } from '@faker-js/faker';
-import { ErrorStrings } from 'ErrorStrings';
+import { CB } from 'vest-utils';
 
-import { dummyTest } from '../../../../testUtils/testDummy';
-
-import { skip, only } from 'focused';
-import { group } from 'group';
-import { TTestSuite } from 'testUtils/TVestMock';
-import { useIsExcluded } from 'useIsExcluded';
 import * as vest from 'vest';
 
-let res: boolean, res1: boolean;
-
-enum FieldNames {
-  f1 = 'f1',
-  f2 = 'f2',
-  f3 = 'f3',
+enum Fields {
+  F1 = 'F1',
+  F2 = 'F2',
+  F3 = 'F3',
+  F4 = 'F4',
 }
 
-describe('focused hooks', () => {
-  test('isExcluded should respect group exclusion', () => {
-    let testObject: vest.IsolateTest;
-    let testObject1: vest.IsolateTest;
+function testSuite(callback: CB) {
+  return vest.staticSuite(callback)();
+}
 
-    const validate = vest.create(() => {
-      testObject = dummyTest.failing();
+describe('Top Level Focus', () => {
+  describe('Top Level Skip', () => {
+    describe('Single field', () => {
+      it('Should skip fields under `skip`', () => {
+        const result = testSuite(() => {
+          vest.skip(Fields.F1);
 
-      group('group_1', () => {
-        vest.skip(true);
-        testObject1 = dummyTest.failing();
-      });
-
-      res = useIsExcluded(testObject);
-      res1 = useIsExcluded(testObject1);
-    });
-
-    validate();
-
-    expect(res).toBe(false);
-    expect(res1).toBe(true);
-  });
-
-  describe('`only` hook', () => {
-    describe('string input', () => {
-      test('isExcluded returns false for included field', () => {
-        vest.create(() => {
-          vest.only(FieldNames.f1);
-          const test = dummyTest.failing(FieldNames.f1);
-          res = useIsExcluded(test);
-        })();
-        expect(res).toBe(false);
-      });
-
-      test('isGroupExcluded returns false for included groups', () => {
-        vest.create(() => {
-          vest.only('group_name');
-
-          group('group_name', () => {
-            res = useIsGroupExcluded('group_name');
-          });
-        })();
-        expect(res).toBe(false);
-      });
-
-      test('isExcluded returns true for non included field', () => {
-        vest.create(() => {
-          vest.only(FieldNames.f1);
-          const t1 = dummyTest.failing(FieldNames.f1);
-          const t2 = dummyTest.failing(FieldNames.f2);
-          res1 = useIsExcluded(t1);
-          res = useIsExcluded(t2);
-        })();
-        expect(res1).toBe(false);
-        expect(res).toBe(true);
-      });
-
-      test('isGroupExcluded returns true for non included group', () => {
-        vest.create(() => {
-          vest.only.group('group_1');
-
-          group('group_1', jest.fn());
-          group('group_2', jest.fn());
-
-          res1 = useIsGroupExcluded('group_1');
-          res = useIsGroupExcluded('group_2');
-        })();
-
-        expect(res1).toBe(false);
-        expect(res).toBe(true);
-      });
-    });
-
-    describe('array input', () => {
-      test('isExcluded returns false for included field', () => {
-        vest.create(() => {
-          vest.only([FieldNames.f1, FieldNames.f2]);
-          const test1 = dummyTest.failing(FieldNames.f1);
-          const test2 = dummyTest.failing(FieldNames.f2);
-          res = useIsExcluded(test1);
-          res1 = useIsExcluded(test2);
-        })();
-        expect(res).toBe(false);
-        expect(res1).toBe(false);
-      });
-
-      test('isGroupExcluded returns false for included groups', () => {
-        const results: boolean[] = [];
-        vest.create(() => {
-          vest.only.group(['group_1', 'group_2']);
-
-          group('group_1', jest.fn());
-          group('group_2', jest.fn());
-          group('group_3', jest.fn());
-
-          results.push(
-            useIsGroupExcluded('group_1'),
-            useIsGroupExcluded('group_2'),
-            useIsGroupExcluded('group_3')
-          );
-        })();
-        expect(results).toEqual([false, false, true]);
-      });
-
-      test('isExcluded returns true for non included field', () => {
-        const results: boolean[] = [];
-        const results1: boolean[] = [];
-        vest.create(() => {
-          const test1 = dummyTest.failing(FieldNames.f1);
-          const test2 = dummyTest.failing(FieldNames.f2);
-          const test3 = dummyTest.failing(FieldNames.f3);
-          results.push(
-            useIsExcluded(test1),
-            useIsExcluded(test2),
-            useIsExcluded(test3)
-          );
-          vest.only([FieldNames.f1, test2.fieldName]);
-          results1.push(
-            useIsExcluded(test1),
-            useIsExcluded(test2),
-            useIsExcluded(test3)
-          );
-        })();
-        expect(results).toEqual([false, false, false]);
-        expect(results1).toEqual([false, false, true]);
-      });
-
-      test('isGroupExcluded returns true for non included groups', () => {
-        const results: boolean[] = [];
-        vest.create(() => {
-          vest.only.group(['group_1', 'group_2']);
-
-          group('group_3', jest.fn());
-          results.push(
-            useIsGroupExcluded('group_1'),
-            useIsGroupExcluded('group_2'),
-            useIsGroupExcluded('group_3')
-          );
-        })();
-        expect(results).toEqual([false, false, true]);
-      });
-    });
-  });
-
-  describe('`skip` hook', () => {
-    describe('string input', () => {
-      test('isExcluded returns true for excluded field', () => {
-        vest.create(() => {
-          vest.skip(FieldNames.f1);
-          const test1 = dummyTest.failing(FieldNames.f1);
-          res = useIsExcluded(test1);
-        })();
-        expect(res).toBe(true);
-      });
-
-      test('isGroupExcluded returns true for excluded groups', () => {
-        vest.create(() => {
-          vest.skip.group('group_1');
-          res = useIsGroupExcluded('group_1');
-          res1 = useIsGroupExcluded('group_2');
-        })();
-
-        expect(res).toBe(true);
-        expect(res1).toBe(false);
-      });
-
-      test('isExcluded returns false for non excluded field', () => {
-        vest.create(() => {
-          vest.skip(FieldNames.f1);
-          res = useIsExcluded(vest.test(FieldNames.f2, jest.fn()));
-        })();
-        expect(res).toBe(false);
-      });
-    });
-
-    test('isGroupExcluded returns false for tests in non excluded groups', () => {
-      vest.create(() => {
-        vest.skip('group_1');
-        res = useIsExcluded(vest.test('field_1', jest.fn()));
-      })();
-      expect(res).toBe(false);
-    });
-
-    describe('array input', () => {
-      test('isExcluded returns true for excluded field', () => {
-        vest.create(() => {
-          vest.skip([FieldNames.f1, FieldNames.f2]);
-          const test1 = dummyTest.failing(FieldNames.f1);
-          const test2 = dummyTest.failing(FieldNames.f2);
-          res = useIsExcluded(test1);
-          res1 = useIsExcluded(test2);
-        })();
-        expect(res).toBe(true);
-        expect(res1).toBe(true);
-      });
-
-      test('isGroupExcluded returns true for excluded groups', () => {
-        const results: boolean[] = [];
-        vest.create(() => {
-          vest.skip.group(['group_1', 'group_2']);
-          results.push(
-            useIsGroupExcluded('group_1'),
-            useIsGroupExcluded('group_2')
-          );
-        })();
-        expect(results).toEqual([true, true]);
-      });
-
-      test('isExcluded returns false for non included field', () => {
-        vest.create(() => {
-          vest.skip([FieldNames.f1, FieldNames.f2]);
-          const test3 = dummyTest.failing(FieldNames.f3);
-          res = useIsExcluded(test3);
-        })();
-        expect(res).toBe(false);
-      });
-
-      test('isGroupExcluded returns false for non excluded groups', () => {
-        vest.create(() => {
-          vest.skip(['group_1', 'group_2']);
-          res = useIsGroupExcluded('group_3');
-        })();
-        expect(res).toBe(false);
-      });
-    });
-
-    describe('Field is in a non included group', () => {
-      let suite: TTestSuite;
-
-      beforeEach(() => {
-        suite = vest.create(() => {
-          vest.only.group('group_1');
-
-          vest.group('group_1', () => {
-            vest.test('field_1', jest.fn());
-          });
-          vest.group('group_2', () => {
-            vest.test('field_2', jest.fn());
-          });
+          vest.test(Fields.F1, 'F1 error', () => false);
+          vest.test(Fields.F2, 'F2 error', () => false);
         });
-        suite();
+        expect(result.hasErrors(Fields.F1)).toBe(false);
+        expect(result.hasErrors(Fields.F2)).toBe(true);
+        expect(result.testCount).toBe(1);
+        expect(result.errorCount).toBe(1);
+        expect(result.tests.F1.errorCount).toBe(0);
+        expect(result.tests.F2.errorCount).toBe(1);
+        expect(result.tests.F1.testCount).toBe(0);
+        expect(result.tests.F2.testCount).toBe(1);
       });
+    });
 
-      it('Should exclude test', () => {
-        expect(suite.get().tests.field_2.testCount).toBe(0);
+    describe('Multiple Fields', () => {
+      it('Should skip fields under `skip`', () => {
+        const result = testSuite(() => {
+          vest.skip([Fields.F1, Fields.F2]);
+
+          vest.test(Fields.F1, 'F1 error', () => false);
+          vest.test(Fields.F2, 'F2 error', () => false);
+          vest.test(Fields.F3, 'F3 error', () => false);
+        });
+        expect(result.hasErrors(Fields.F1)).toBe(false);
+        expect(result.hasErrors(Fields.F2)).toBe(false);
+        expect(result.hasErrors(Fields.F3)).toBe(true);
+        expect(result.testCount).toBe(1);
+        expect(result.errorCount).toBe(1);
+        expect(result.tests.F1.errorCount).toBe(0);
+        expect(result.tests.F2.errorCount).toBe(0);
+        expect(result.tests.F3.errorCount).toBe(1);
+        expect(result.tests.F1.testCount).toBe(0);
+        expect(result.tests.F2.testCount).toBe(0);
+        expect(result.tests.F3.testCount).toBe(1);
+      });
+    });
+
+    describe('Multiple Skip Calls', () => {
+      it('Should skip fields under `skip`', () => {
+        const result = testSuite(() => {
+          vest.skip(Fields.F1);
+          vest.skip(Fields.F2);
+
+          vest.test(Fields.F1, 'F1 error', () => false);
+          vest.test(Fields.F2, 'F2 error', () => false);
+          vest.test(Fields.F3, 'F3 error', () => false);
+        });
+        expect(result.hasErrors(Fields.F1)).toBe(false);
+        expect(result.hasErrors(Fields.F2)).toBe(false);
+        expect(result.hasErrors(Fields.F3)).toBe(true);
+        expect(result.testCount).toBe(1);
+        expect(result.errorCount).toBe(1);
+        expect(result.tests.F1.errorCount).toBe(0);
+        expect(result.tests.F2.errorCount).toBe(0);
+        expect(result.tests.F3.errorCount).toBe(1);
+        expect(result.tests.F1.testCount).toBe(0);
+        expect(result.tests.F2.testCount).toBe(0);
+        expect(result.tests.F3.testCount).toBe(1);
+      });
+    });
+
+    describe('Nonexistent Field', () => {
+      it('Should ignore nonexistent fields', () => {
+        const result = testSuite(() => {
+          vest.skip(Fields.F1);
+
+          vest.test(Fields.F2, 'F2 error', () => false);
+          vest.test(Fields.F3, 'F3 error', () => false);
+        });
+        expect(result.hasErrors(Fields.F1)).toBe(false);
+        expect(result.hasErrors(Fields.F2)).toBe(true);
+        expect(result.hasErrors(Fields.F3)).toBe(true);
+        expect(result.testCount).toBe(2);
+        expect(result.errorCount).toBe(2);
+        expect(result.tests.F1).toBeUndefined();
+        expect(result.tests.F2.errorCount).toBe(1);
+        expect(result.tests.F3.errorCount).toBe(1);
+        expect(result.tests.F2.testCount).toBe(1);
+        expect(result.tests.F3.testCount).toBe(1);
+      });
+    });
+
+    describe('Empty skip call', () => {
+      it('Should ignore skip call', () => {
+        const result = testSuite(() => {
+          vest.skip();
+
+          vest.test(Fields.F1, 'F1 error', () => false);
+          vest.test(Fields.F2, 'F2 error', () => false);
+        });
+        expect(result.hasErrors(Fields.F1)).toBe(true);
+        expect(result.hasErrors(Fields.F2)).toBe(true);
+        expect(result.testCount).toBe(2);
+        expect(result.errorCount).toBe(2);
+        expect(result.tests.F1.errorCount).toBe(1);
+        expect(result.tests.F2.errorCount).toBe(1);
+        expect(result.tests.F1.testCount).toBe(1);
+        expect(result.tests.F2.testCount).toBe(1);
       });
     });
   });
 
-  describe('Error handling', () => {
-    describe('When called outside of a suite', () => {
-      it('Should throw an error', () => {
-        expect(() => only(faker.random.word())).toThrow(
-          ErrorStrings.HOOK_CALLED_OUTSIDE
-        );
-        expect(() => skip(faker.random.word())).toThrow(
-          ErrorStrings.HOOK_CALLED_OUTSIDE
-        );
+  describe('Top Level Only', () => {
+    describe('Single field', () => {
+      it('Should only run fields under `only`', () => {
+        const result = testSuite(() => {
+          vest.only(Fields.F1);
+
+          vest.test(Fields.F1, 'F1 error', () => false);
+          vest.test(Fields.F2, 'F2 error', () => false);
+        });
+        expect(result.hasErrors(Fields.F1)).toBe(true);
+        expect(result.hasErrors(Fields.F2)).toBe(false);
+        expect(result.testCount).toBe(1);
+        expect(result.errorCount).toBe(1);
+        expect(result.tests.F1.errorCount).toBe(1);
+        expect(result.tests.F2.errorCount).toBe(0);
+        expect(result.tests.F1.testCount).toBe(1);
+        expect(result.tests.F2.testCount).toBe(0);
+      });
+    });
+
+    describe('Multiple Fields', () => {
+      it('Should only run fields under `only`', () => {
+        const result = testSuite(() => {
+          vest.only([Fields.F1, Fields.F2]);
+
+          vest.test(Fields.F1, 'F1 error', () => false);
+          vest.test(Fields.F2, 'F2 error', () => false);
+          vest.test(Fields.F3, 'F3 error', () => false);
+        });
+        expect(result.hasErrors(Fields.F1)).toBe(true);
+        expect(result.hasErrors(Fields.F2)).toBe(true);
+        expect(result.hasErrors(Fields.F3)).toBe(false);
+        expect(result.testCount).toBe(2);
+        expect(result.errorCount).toBe(2);
+        expect(result.tests.F1.errorCount).toBe(1);
+        expect(result.tests.F2.errorCount).toBe(1);
+        expect(result.tests.F3.errorCount).toBe(0);
+        expect(result.tests.F1.testCount).toBe(1);
+        expect(result.tests.F2.testCount).toBe(1);
+        expect(result.tests.F3.testCount).toBe(0);
+      });
+    });
+
+    describe('Multiple Only Calls', () => {
+      it('Should only run fields under `only`', () => {
+        const result = testSuite(() => {
+          vest.only(Fields.F1);
+          vest.only(Fields.F2);
+
+          vest.test(Fields.F1, 'F1 error', () => false);
+          vest.test(Fields.F2, 'F2 error', () => false);
+          vest.test(Fields.F3, 'F3 error', () => false);
+        });
+        expect(result.hasErrors(Fields.F1)).toBe(true);
+        expect(result.hasErrors(Fields.F2)).toBe(true);
+        expect(result.hasErrors(Fields.F3)).toBe(false);
+        expect(result.testCount).toBe(2);
+        expect(result.errorCount).toBe(2);
+        expect(result.tests.F1.errorCount).toBe(1);
+        expect(result.tests.F2.errorCount).toBe(1);
+        expect(result.tests.F3.errorCount).toBe(0);
+        expect(result.tests.F1.testCount).toBe(1);
+        expect(result.tests.F2.testCount).toBe(1);
+        expect(result.tests.F3.testCount).toBe(0);
+      });
+    });
+
+    describe('Nonexistent Field', () => {
+      it('Should ignore all other fields', () => {
+        const result = testSuite(() => {
+          vest.only(Fields.F1);
+
+          vest.test(Fields.F2, 'F2 error', () => false);
+          vest.test(Fields.F3, 'F3 error', () => false);
+        });
+        expect(result.hasErrors(Fields.F1)).toBe(false);
+        expect(result.hasErrors(Fields.F2)).toBe(false);
+        expect(result.hasErrors(Fields.F3)).toBe(false);
+
+        expect(result.testCount).toBe(0);
+        expect(result.errorCount).toBe(0);
+        expect(result.tests.F1).toBeUndefined();
+        expect(result.tests.F2.testCount).toBe(0);
+        expect(result.tests.F3.testCount).toBe(0);
       });
     });
   });
 });
 
-// FIXME: These are lower quality tests, and we need to replace them with more robust ones
-// describe('isExcluded', () => {
-//   let exclusion: Partial<TExclusion>;
+describe('Scoped Focus', () => {
+  describe('Only', () => {
+    it('Should apply only to the current scope', () => {
+      const result = testSuite(() => {
+        vest.group('Group', () => {
+          vest.only(Fields.F2);
 
-//   const runIsExcluded = (
-//     exclusion: Partial<TExclusion>,
-//     testObject: IsolateTest
-//   ) => {
-//     vest.staticSuite(() => {
+          vest.test(Fields.F1, 'F1 error', () => false);
+          vest.test(Fields.F2, 'F2 error', () => false);
+          vest.test(Fields.F3, 'F3 error', () => false);
+        });
 
-//     })
-//     // SuiteContext.run({}, () => {
-//     //   Object.assign(useExclusion(), exclusion);
-//     //   const res = useIsExcluded(testObject);
+        vest.test(Fields.F4, 'F4 error', () => false);
+      });
 
-//     //   return res;
-//     // });
-//   }
+      expect(result.hasErrors(Fields.F1)).toBe(false);
+      expect(result.hasErrors(Fields.F2)).toBe(true);
+      expect(result.hasErrors(Fields.F3)).toBe(false);
+      expect(result.hasErrors(Fields.F4)).toBe(true);
+      expect(result.tests.F1.testCount).toBe(0);
+      expect(result.tests.F2.testCount).toBe(1);
+      expect(result.tests.F3.testCount).toBe(0);
+    });
 
-//   const genTest = (fieldName: string, groupName?: string) =>
-//     new IsolateTest({
-//       fieldName,
-//       testFn: jest.fn(),
-//       groupName,
-//     });
-//   describe('skip', () => {
-//     beforeEach(() => {
-//       exclusion = { tests: { field_1: false, field_2: false } };
-//     });
-//     it('returns true for skipped field', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_1'))).toBe(true);
-//       expect(runIsExcluded(exclusion, genTest('field_2', 'group_1'))).toBe(
-//         true
-//       );
-//     });
-//     it('returns false for non skipped field', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_3'))).toBe(false);
-//       expect(runIsExcluded(exclusion, genTest('field_4', 'group_1'))).toBe(
-//         false
-//       );
-//     });
-//   });
-//   describe('only', () => {
-//     beforeEach(() => {
-//       exclusion = { tests: { field_1: true, field_2: true } };
-//     });
-//     it('returns false for included field', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_1'))).toBe(false);
-//       expect(runIsExcluded(exclusion, genTest('field_2', 'group_1'))).toBe(
-//         false
-//       );
-//     });
-//     it('returns true for non included field', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_3'))).toBe(true);
-//       expect(runIsExcluded(exclusion, genTest('field_4', 'group_1'))).toBe(
-//         true
-//       );
-//     });
-//   });
-//   describe('only+skip', () => {
-//     beforeEach(() => {
-//       exclusion = {
-//         tests: { field_1: true, field_2: true, field_3: false, field_4: false },
-//       };
-//     });
-//     it('returns false for included tests', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_1'))).toBe(false);
-//       expect(runIsExcluded(exclusion, genTest('field_2', 'group_1'))).toBe(
-//         false
-//       );
-//     });
-//     it('returns true excluded tests', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_3'))).toBe(true);
-//       expect(runIsExcluded(exclusion, genTest('field_4', 'group_1'))).toBe(
-//         true
-//       );
-//     });
-//     it('returns true for non included field', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_5'))).toBe(true);
-//       expect(runIsExcluded(exclusion, genTest('field_6', 'group_1'))).toBe(
-//         true
-//       );
-//     });
-//   });
-//   describe('skip.group', () => {
-//     beforeEach(() => {
-//       exclusion = {
-//         groups: {
-//           group_1: false,
-//           group_2: false,
-//         },
-//       };
-//     });
+    describe('When there are tests of the same field outside of the scope', () => {
+      it('Should only apply within the scope', () => {
+        const result = testSuite(() => {
+          vest.mode(vest.Modes.ALL); // This shows us that the test actually run outside of the scope
+          vest.group('Group', () => {
+            vest.only(Fields.F2);
 
-//     it('Returns true for tests in skipped group', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_1', 'group_1'))).toBe(
-//         true
-//       );
-//       expect(runIsExcluded(exclusion, genTest('field_2', 'group_2'))).toBe(
-//         true
-//       );
-//     });
-//     it('Returns false for tests in non skipped groups', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_3', 'group_3'))).toBe(
-//         false
-//       );
-//       expect(runIsExcluded(exclusion, genTest('field_4', 'group_4'))).toBe(
-//         false
-//       );
-//     });
-//     it('Returns false for tests outside of any group', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_3'))).toBe(false);
-//       expect(runIsExcluded(exclusion, genTest('field_4'))).toBe(false);
-//     });
-//   });
-//   describe('only.group', () => {
-//     beforeEach(() => {
-//       exclusion = {
-//         groups: {
-//           group_1: true,
-//           group_2: true,
-//         },
-//       };
-//     });
+            vest.test(Fields.F1, 'F1 error', () => false);
+            vest.test(Fields.F2, 'F2 error', () => false);
+            vest.test(Fields.F3, 'F3 error', () => false);
+          });
 
-//     it('returns false for tests in included groups', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_1', 'group_1'))).toBe(
-//         false
-//       );
-//       expect(runIsExcluded(exclusion, genTest('field_1', 'group_2'))).toBe(
-//         false
-//       );
-//     });
+          vest.test(Fields.F1, 'F1 2nd error', () => false);
+          vest.test(Fields.F2, 'F2 2nd error', () => false);
+          vest.test(Fields.F3, 'F3 2nd error', () => false);
+        });
+        expect(result.hasErrors(Fields.F1)).toBe(true);
+        expect(result.hasErrors(Fields.F2)).toBe(true);
+        expect(result.hasErrors(Fields.F3)).toBe(true);
+        expect(result.tests.F1.testCount).toBe(1);
+        expect(result.tests.F2.testCount).toBe(2);
+        expect(result.tests.F3.testCount).toBe(1);
+      });
+    });
+  });
 
-//     it('returns true for groups in non included groups', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_1', 'group_3'))).toBe(
-//         true
-//       );
-//       expect(runIsExcluded(exclusion, genTest('field_2', 'group_4'))).toBe(
-//         true
-//       );
-//     });
+  describe('Skip', () => {
+    it('Should apply only to the current scope', () => {
+      const result = testSuite(() => {
+        vest.group('Group', () => {
+          vest.skip(Fields.F2);
 
-//     it('returns true for tests outside of any group', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_1'))).toBe(true);
-//       expect(runIsExcluded(exclusion, genTest('field_2'))).toBe(true);
-//     });
-//   });
+          vest.test(Fields.F1, 'F1 error', () => false);
+          vest.test(Fields.F2, 'F2 error', () => false);
+          vest.test(Fields.F3, 'F3 error', () => false);
+        });
 
-//   describe('only.group + only', () => {
-//     beforeEach(() => {
-//       exclusion = {
-//         groups: {
-//           group_1: true,
-//           group_2: true,
-//         },
-//         tests: { field_1: true, field_2: true },
-//       };
-//     });
+        vest.test(Fields.F4, 'F4 error', () => false);
+      });
 
-//     it('returns true for included tests outside of the group', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_1'))).toBe(true);
-//       expect(runIsExcluded(exclusion, genTest('field_2'))).toBe(true);
-//     });
+      expect(result.hasErrors(Fields.F1)).toBe(true);
+      expect(result.hasErrors(Fields.F2)).toBe(false);
+      expect(result.hasErrors(Fields.F3)).toBe(true);
+      expect(result.hasErrors(Fields.F4)).toBe(true);
+      expect(result.tests.F1.testCount).toBe(1);
+      expect(result.tests.F2.testCount).toBe(0);
+      expect(result.tests.F3.testCount).toBe(1);
+      expect(result.tests.F4.testCount).toBe(1);
+    });
 
-//     it('returns false for included tests in included groups', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_1', 'group_1'))).toBe(
-//         false
-//       );
-//       expect(runIsExcluded(exclusion, genTest('field_2', 'group_2'))).toBe(
-//         false
-//       );
-//     });
+    describe('When there are tests of the same field outside of the scope', () => {
+      it('Should only apply within the scope', () => {
+        const result = testSuite(() => {
+          vest.mode(vest.Modes.ALL); // This shows us that the tests actually run outside of the scope
+          vest.group('Group', () => {
+            vest.skip(Fields.F2);
 
-//     it('returns true for included test in non included group', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_1', 'group_3'))).toBe(
-//         true
-//       );
-//       expect(runIsExcluded(exclusion, genTest('field_2', 'group_4'))).toBe(
-//         true
-//       );
-//     });
+            vest.test(Fields.F1, 'F1 error', () => false);
+            vest.test(Fields.F2, 'F2 error', () => false);
+            vest.test(Fields.F3, 'F3 error', () => false);
+          });
 
-//     it('returns true for non included test in included group', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_3', 'group_1'))).toBe(
-//         true
-//       );
-//       expect(runIsExcluded(exclusion, genTest('field_4', 'group_2'))).toBe(
-//         true
-//       );
-//     });
+          vest.test(Fields.F1, 'F1 2nd error', () => false);
+          vest.test(Fields.F2, 'F2 2nd error', () => false);
+          vest.test(Fields.F3, 'F3 2nd error', () => false);
+        });
+        expect(result.hasErrors(Fields.F1)).toBe(true);
+        expect(result.hasErrors(Fields.F2)).toBe(true);
+        expect(result.hasErrors(Fields.F3)).toBe(true);
+        expect(result.tests.F1.testCount).toBe(2);
+        expect(result.tests.F2.testCount).toBe(1);
+        expect(result.tests.F3.testCount).toBe(2);
+      });
+    });
+  });
 
-//     it('returns true for non included tests', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_3'))).toBe(true);
-//       expect(runIsExcluded(exclusion, genTest('field_4'))).toBe(true);
-//     });
-//   });
+  describe('Only and Skip', () => {
+    it('Should apply only to the current scope', () => {
+      const result = testSuite(() => {
+        vest.group('Group', () => {
+          vest.only(Fields.F2);
+          vest.skip(Fields.F3);
 
-//   describe('skip.group + only', () => {
-//     beforeEach(() => {
-//       exclusion = {
-//         groups: {
-//           group_1: false,
-//           group_2: false,
-//         },
-//         tests: { field_1: true, field_2: true },
-//       };
-//     });
+          vest.test(Fields.F1, 'F1 error', () => false);
+          vest.test(Fields.F2, 'F2 error', () => false);
+          vest.test(Fields.F3, 'F3 error', () => false);
+        });
 
-//     it('returns true for included tests in excluded groups', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_1', 'group_1'))).toBe(
-//         true
-//       );
-//       expect(runIsExcluded(exclusion, genTest('field_2', 'group_1'))).toBe(
-//         true
-//       );
-//       expect(runIsExcluded(exclusion, genTest('field_3', 'group_1'))).toBe(
-//         true
-//       );
-//       expect(runIsExcluded(exclusion, genTest('field_1', 'group_2'))).toBe(
-//         true
-//       );
-//       expect(runIsExcluded(exclusion, genTest('field_2', 'group_2'))).toBe(
-//         true
-//       );
-//       expect(runIsExcluded(exclusion, genTest('field_3', 'group_2'))).toBe(
-//         true
-//       );
-//     });
+        vest.test(Fields.F4, 'F4 error', () => false);
+      });
 
-//     it('returns false for included tests in non excluded groups', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_1', 'group_3'))).toBe(
-//         false
-//       );
-//       expect(runIsExcluded(exclusion, genTest('field_2', 'group_3'))).toBe(
-//         false
-//       );
-//       expect(runIsExcluded(exclusion, genTest('field_1', 'group_4'))).toBe(
-//         false
-//       );
-//       expect(runIsExcluded(exclusion, genTest('field_2', 'group_4'))).toBe(
-//         false
-//       );
-//     });
+      expect(result.hasErrors(Fields.F1)).toBe(false);
+      expect(result.hasErrors(Fields.F2)).toBe(true);
+      expect(result.hasErrors(Fields.F3)).toBe(false);
+      expect(result.hasErrors(Fields.F4)).toBe(true);
+      expect(result.tests.F1.testCount).toBe(0);
+      expect(result.tests.F2.testCount).toBe(1);
+      expect(result.tests.F3.testCount).toBe(0);
+      expect(result.tests.F4.testCount).toBe(1);
+    });
 
-//     it('returns true for non included tests', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_3'))).toBe(true);
-//       expect(runIsExcluded(exclusion, genTest('field_4'))).toBe(true);
-//     });
+    describe('When there are tests of the same field outside of the scope', () => {
+      it('Should only apply within the scope', () => {
+        const result = testSuite(() => {
+          vest.mode(vest.Modes.ALL); // This shows us that the tests actually run outside of the scope
+          vest.group('Group', () => {
+            vest.only(Fields.F2);
+            vest.skip(Fields.F3);
 
-//     it('returns false for included tests outside of the group', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_1'))).toBe(false);
-//       expect(runIsExcluded(exclusion, genTest('field_2'))).toBe(false);
-//     });
-//   });
+            vest.test(Fields.F1, 'F1 error', () => false);
+            vest.test(Fields.F2, 'F2 error', () => false);
+            vest.test(Fields.F3, 'F3 error', () => false);
+          });
 
-//   describe('only.group + skip', () => {
-//     beforeEach(() => {
-//       exclusion = {
-//         groups: {
-//           group_1: true,
-//           group_2: true,
-//         },
-//         tests: { field_1: false, field_2: false },
-//       };
-//     });
+          vest.test(Fields.F1, 'F1 2nd error', () => false);
+          vest.test(Fields.F2, 'F2 2nd error', () => false);
+          vest.test(Fields.F3, 'F3 2nd error', () => false);
+        });
+        expect(result.hasErrors(Fields.F1)).toBe(true);
+        expect(result.hasErrors(Fields.F2)).toBe(true);
+        expect(result.hasErrors(Fields.F3)).toBe(true);
+        expect(result.tests.F1.testCount).toBe(1);
+        expect(result.tests.F2.testCount).toBe(2);
+        expect(result.tests.F3.testCount).toBe(1);
+      });
+    });
+  });
+});
 
-//     it('returns true for excluded tests', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_1'))).toBe(true);
-//       expect(runIsExcluded(exclusion, genTest('field_2'))).toBe(true);
-//     });
-//     it('returns true for excluded test in included group', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_1', 'group_1'))).toBe(
-//         true
-//       );
-//       expect(runIsExcluded(exclusion, genTest('field_2', 'group_2'))).toBe(
-//         true
-//       );
-//     });
-//     it('returns true for excluded test in non included group', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_1', 'group_3'))).toBe(
-//         true
-//       );
-//       expect(runIsExcluded(exclusion, genTest('field_2', 'group_4'))).toBe(
-//         true
-//       );
-//     });
-//     it('returns false for non excluded test in included group', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_3', 'group_1'))).toBe(
-//         false
-//       );
-//       expect(runIsExcluded(exclusion, genTest('field_4', 'group_2'))).toBe(
-//         false
-//       );
-//     });
-//     it('returns true for non excluded test in non included group', () => {
-//       expect(runIsExcluded(exclusion, genTest('field_3', 'group_3'))).toBe(
-//         true
-//       );
-//       expect(runIsExcluded(exclusion, genTest('field_4', 'group_4'))).toBe(
-//         true
-//       );
-//     });
-//   });
+describe('Multiple scopes', () => {
+  describe('Top and lower level scopes', () => {
+    describe('only>skip', () => {
+      it('Should include the field globally but skip in the scope', () => {
+        const result = testSuite(() => {
+          vest.mode(vest.Modes.ALL); // This shows us that the tests actually run outside of the scope
 
-//   describe('Invalid input', () => {
-//     describe('ExclusionItem is not a string', () => {
-//       it('Should skip registration', () => {
-//         const suite = vest.create('suite', () => {
-//           // @ts-ignore - Invalid input
-//           vest.only(111);
+          vest.only(Fields.F1);
+          vest.test(Fields.F1, 'F1 1st error', () => false);
+          vest.test(Fields.F2, 'F2 1st error', () => false);
+          vest.group('Group', () => {
+            vest.skip(Fields.F1);
+            vest.test(Fields.F1, 'F1 2nd error', () => false);
+            vest.test(Fields.F2, 'F2 2nd error', () => false);
+          });
+          vest.test(Fields.F1, 'F1 3rd error', () => false);
+          vest.test(Fields.F2, 'F2 3rd error', () => false);
+        });
+        expect(result.hasErrors(Fields.F1)).toBe(true);
+        expect(result.hasErrors(Fields.F2)).toBe(false);
+        expect(result.tests.F1.testCount).toBe(2);
+        expect(result.tests.F1.testCount).toBe(2);
+        expect(result.groups.Group.F1.testCount).toBe(0);
+        expect(result.tests.F2.testCount).toBe(0);
+        expect(result.tests.F2.testCount).toBe(0);
+      });
+    });
 
-//           vest.test('f1', () => false);
-//         });
+    describe('skip>only', () => {
+      it('Should skip the field globally and include it in the scope', () => {
+        const result = testSuite(() => {
+          vest.mode(vest.Modes.ALL); // This shows us that the tests actually run outside of the scope
 
-//         expect(suite().hasErrors('f1')).toBe(true);
-//       });
-//     });
-//   });
-// });
+          vest.skip(Fields.F1);
+          vest.test(Fields.F1, 'F1 1st error', () => false);
+          vest.test(Fields.F2, 'F2 1st error', () => false);
+          vest.group('Group', () => {
+            vest.only(Fields.F1);
+            vest.test(Fields.F1, 'F1 2nd error', () => false);
+            vest.test(Fields.F2, 'F2 2nd error', () => false);
+          });
+          vest.test(Fields.F1, 'F1 3rd error', () => false);
+          vest.test(Fields.F2, 'F2 3rd error', () => false);
+        });
+        expect(result.hasErrors(Fields.F1)).toBe(true);
+        expect(result.hasErrors(Fields.F2)).toBe(true);
+        expect(result.tests.F1.testCount).toBe(1);
+        expect(result.groups.Group.F1.testCount).toBe(1);
+        expect(result.tests.F2.testCount).toBe(2);
+        expect(result.groups.Group.F2.testCount).toBe(0);
+      });
+    });
+  });
+});
+
+describe('skip(true)', () => {
+  it('Should skip all tests within the current scope and its children', () => {
+    const result = testSuite(() => {
+      vest.mode(vest.Modes.ALL); // This shows us that the tests actually run outside of the scope
+
+      vest.skip(true);
+      vest.test(Fields.F1, 'F1 1st error', () => false);
+      vest.test(Fields.F2, 'F2 1st error', () => false);
+      vest.group('Group', () => {
+        vest.test(Fields.F1, 'F1 2nd error', () => false);
+        vest.test(Fields.F2, 'F2 2nd error', () => false);
+      });
+      vest.test(Fields.F1, 'F1 3rd error', () => false);
+      vest.test(Fields.F2, 'F2 3rd error', () => false);
+    });
+    expect(result.hasErrors(Fields.F1)).toBe(false);
+    expect(result.hasErrors(Fields.F2)).toBe(false);
+    expect(result.tests.F1.testCount).toBe(0);
+    expect(result.tests.F2.testCount).toBe(0);
+    expect(result.groups.Group.F1.testCount).toBe(0);
+    expect(result.groups.Group.F2.testCount).toBe(0);
+  });
+
+  it('Should only apply to the current scope', () => {
+    const result = testSuite(() => {
+      vest.mode(vest.Modes.ALL); // This shows us that the tests actually run outside of the scope
+
+      vest.test(Fields.F1, 'F1 1st error', () => false);
+      vest.test(Fields.F2, 'F2 1st error', () => false);
+      vest.group('Group', () => {
+        vest.skip(true);
+        vest.test(Fields.F1, 'F1 2nd error', () => false);
+        vest.test(Fields.F2, 'F2 2nd error', () => false);
+      });
+      vest.test(Fields.F1, 'F1 3rd error', () => false);
+      vest.test(Fields.F2, 'F2 3rd error', () => false);
+    });
+    expect(result.hasErrors(Fields.F1)).toBe(true);
+    expect(result.hasErrors(Fields.F2)).toBe(true);
+    expect(result.tests.F1.testCount).toBe(2);
+    expect(result.tests.F2.testCount).toBe(2);
+    expect(result.groups.Group.F1.testCount).toBe(0);
+    expect(result.groups.Group.F2.testCount).toBe(0);
+  });
+});
